refactor: move models and utility functions into individual files
