+++ src/rds-control-plane-mcp-server/awslabs/rds_control_plane_mcp_server/resources/db_cluster/get_cluster_detail.py
@@ -17,11 +17,120 @@
 import asyncio
 from ...common.connection import RDSConnectionManager
 from ...common.decorator import handle_exceptions
-from ...common.models import ClusterModel
 from ...common.server import mcp
-from .utils import format_cluster_detail
+from ...common.utils import convert_datetime_to_string
 from loguru import logger
-from pydantic import Field
+from mypy_boto3_rds.type_defs import DBClusterTypeDef
+from pydantic import BaseModel, Field
+from typing import Dict, List, Optional
+
+
+# Data Models
+
+
+class ClusterMember(BaseModel):
+    """DB cluster member model."""
+
+    instance_id: str = Field(description='The instance identifier of the DB cluster member')
+    is_writer: bool = Field(description='Whether the cluster member is a writer instance')
+    status: Optional[str] = Field(
+        None, description='The status of the DB cluster parameter group for this member'
+    )
+
+
+class Cluster(BaseModel):
+    """DB cluster model."""
+
+    cluster_id: str = Field(description='The DB cluster identifier')
+    status: str = Field(description='The current status of the DB cluster')
+    engine: str = Field(description='The database engine')
+    engine_version: Optional[str] = Field(None, description='The version of the database engine')
+    endpoint: Optional[str] = Field(
+        None, description='The connection endpoint for the primary instance'
+    )
+    reader_endpoint: Optional[str] = Field(
+        None, description='The reader endpoint for the DB cluster'
+    )
+    multi_az: bool = Field(
+        description='Whether the DB cluster has instances in multiple Availability Zones'
+    )
+    backup_retention: int = Field(description='The retention period for automated backups')
+    preferred_backup_window: Optional[str] = Field(
+        None, description='The daily time range during which automated backups are created'
+    )
+    preferred_maintenance_window: Optional[str] = Field(
+        None, description='The weekly time range during which system maintenance can occur'
+    )
+    created_time: Optional[str] = Field(
+        None, description='The time when the DB cluster was created'
+    )
+    members: List[ClusterMember] = Field(
+        default_factory=list, description='A list of DB cluster members'
+    )
+    vpc_security_groups: List[Dict[str, str]] = Field(
+        default_factory=list, description='A list of VPC security groups the DB cluster belongs to'
+    )
+    tags: Dict[str, str] = Field(default_factory=dict, description='A list of tags')
+    resource_uri: Optional[str] = Field(None, description='The resource URI for this cluster')
+
+
+# Helper Functions
+
+
+def format_cluster_detail(cluster: DBClusterTypeDef) -> Cluster:
+    """Format cluster information from AWS API response into a detailed structured model.
+
+    This method transforms the raw AWS API response data into a standardized
+    Cluster object, extracting and organizing key cluster attributes
+    including members, security groups, and tags.
+
+    Args:
+        cluster: Raw cluster data from AWS API response
+
+    Returns:
+        Formatted cluster information as a Cluster object with comprehensive details
+    """
+    members = []
+    for member in cluster.get('DBClusterMembers', []):
+        members.append(
+            ClusterMember(
+                instance_id=member.get('DBInstanceIdentifier', ''),
+                is_writer=member.get('IsClusterWriter', False),
+                status=member.get('DBClusterParameterGroupStatus'),
+            )
+        )
+
+    vpc_security_groups = []
+    for sg in cluster.get('VpcSecurityGroups', []):
+        vpc_security_groups.append(
+            {'id': sg.get('VpcSecurityGroupId', ''), 'status': sg.get('Status', '')}
+        )
+
+    tags = {}
+    if cluster.get('TagList'):
+        for tag in cluster.get('TagList', []):
+            if 'Key' in tag and 'Value' in tag:
+                tags[tag['Key']] = tag['Value']
+
+    cluster_id = cluster.get('DBClusterIdentifier', '')
+
+    return Cluster(
+        cluster_id=cluster_id,
+        status=cluster.get('Status', ''),
+        engine=cluster.get('Engine', ''),
+        engine_version=cluster.get('EngineVersion'),
+        endpoint=cluster.get('Endpoint'),
+        reader_endpoint=cluster.get('ReaderEndpoint'),
+        multi_az=cluster.get('MultiAZ', False),
+        backup_retention=cluster.get('BackupRetentionPeriod', 0),
+        preferred_backup_window=cluster.get('PreferredBackupWindow'),
+        preferred_maintenance_window=cluster.get('PreferredMaintenanceWindow'),
+        created_time=convert_datetime_to_string(cluster.get('ClusterCreateTime')),
+        members=members,
+        vpc_security_groups=vpc_security_groups,
+        tags=tags,
+        resource_uri=f'aws-rds://db-cluster/{cluster_id}',
+    )
 
 
 GET_CLUSTER_DETAIL_RESOURCE_DESCRIPTION = """Get detailed information about a specific Amazon RDS cluster.
@@ -65,21 +174,14 @@ async def get_cluster_detail(
     cluster_id: str = Field(
         ..., description='The unique identifier of the RDS DB cluster to retrieve details for'
     ),
-) -> ClusterModel:
+) -> Cluster:
     """Retrieve detailed information about a specific RDS cluster.
-
-    This method queries the AWS RDS API for comprehensive information about
-    a specific DB cluster identified by its unique cluster ID. It handles
-    AWS API interactions, error conditions, and formats the response data
-    into a consistent JSON structure.
 
     Args:
         cluster_id: The unique identifier of the DB cluster to retrieve
 
     Returns:
-        JSON string containing detailed cluster information including configuration,
-        status, endpoints, backup settings, member instances, and security groups.
-        Returns error information if the cluster doesn't exist or cannot be accessed.
+        Formatted Cluster object with comprehensive details
     """
     logger.info(f'Getting cluster detail resource for {cluster_id}')
     rds_client = RDSConnectionManager.get_connection()
@@ -91,6 +193,5 @@ async def get_cluster_detail(
     if not clusters:
         raise ValueError(f'Cluster {cluster_id} not found')
     cluster = format_cluster_detail(clusters[0])
-    cluster.resource_uri = f'aws-rds://db-cluster/{cluster_id}'
 
     return cluster
