+++ src/rds-control-plane-mcp-server/awslabs/rds_control_plane_mcp_server/resources/db_instance/get_instance_detail.py
@@ -17,67 +17,175 @@
 import asyncio
 from ...common.connection import RDSConnectionManager
 from ...common.decorator import handle_exceptions
-from ...common.models import InstanceModel
-from ...common.server import mcp
-from .utils import format_instance_detail
 from loguru import logger
-from pydantic import Field
+from mypy_boto3_rds.type_defs import DBInstanceTypeDef
+from pydantic import BaseModel, Field
+from typing import Dict, List, Optional
+
+
+class InstanceEndpoint(BaseModel):
+    """DB instance endpoint model."""
+
+    address: Optional[str] = Field(None, description='The DNS address of the instance')
+    port: Optional[int] = Field(
+        None, description='The port that the database engine is listening on'
+    )
+    hosted_zone_id: Optional[str] = Field(
+        None, description='The ID of the Amazon Route 53 hosted zone'
+    )
+
+
+class InstanceStorage(BaseModel):
+    """DB instance storage model."""
+
+    type: Optional[str] = Field(None, description='The storage type')
+    allocated: Optional[int] = Field(None, description='The allocated storage size in gibibytes')
+    encrypted: Optional[bool] = Field(None, description='Whether the storage is encrypted')
+
+
+class Instance(BaseModel):
+    """DB instance model."""
+
+    instance_id: str = Field(description='The DB instance identifier')
+    status: str = Field(description='The current status of the DB instance')
+    engine: str = Field(description='The database engine')
+    engine_version: Optional[str] = Field(None, description='The version of the database engine')
+    instance_class: str = Field(
+        description='The compute and memory capacity class of the DB instance'
+    )
+    endpoint: Optional[InstanceEndpoint] = Field(None, description='The connection endpoint')
+    availability_zone: Optional[str] = Field(
+        None, description='The Availability Zone of the DB instance'
+    )
+    multi_az: bool = Field(description='Whether the DB instance is a Multi-AZ deployment')
+    storage: Optional[InstanceStorage] = Field(None, description='The storage configuration')
+    preferred_backup_window: Optional[str] = Field(
+        None, description='The daily time range during which automated backups are created'
+    )
+    preferred_maintenance_window: Optional[str] = Field(
+        None, description='The weekly time range during which system maintenance can occur'
+    )
+    publicly_accessible: bool = Field(description='Whether the DB instance is publicly accessible')
+    vpc_security_groups: List[Dict[str, str]] = Field(
+        default_factory=list,
+        description='A list of VPC security groups the DB instance belongs to',
+    )
+    db_cluster: Optional[str] = Field(
+        None, description='The DB cluster identifier, if this is a member of a DB cluster'
+    )
+    tags: Dict[str, str] = Field(default_factory=dict, description='A list of tags')
+    dbi_resource_id: Optional[str] = Field(
+        None, description='The AWS Region-unique, immutable identifier for the DB instance'
+    )
+    resource_uri: Optional[str] = Field(None, description='The resource URI for this instance')
+
+
+def format_instance_detail(instance: DBInstanceTypeDef) -> Instance:
+    """Format instance information into a detailed model with comprehensive information.
+
+    Takes raw instance data from AWS and formats it into a structured Instance model
+    containing all relevant details about the RDS instance configuration.
+
+    Args:
+        instance: Raw instance data from AWS
+
+    Returns:
+        Instance: Formatted instance information with comprehensive details
+    """
+    endpoint = InstanceEndpoint(
+        address=instance.get('Endpoint', {}).get('Address'),
+        hosted_zone_id=instance.get('Endpoint', {}).get('HostedZoneId'),
+        port=instance.get('Endpoint', {}).get('Port'),
+    )
+
+    storage = InstanceStorage(
+        type=instance.get('StorageType'),
+        allocated=instance.get('AllocatedStorage'),
+        encrypted=instance.get('StorageEncrypted'),
+    )
+
+    vpc_security_groups = []
+    for sg in instance.get('VpcSecurityGroups', []):
+        vpc_security_groups.append(
+            {'id': sg.get('VpcSecurityGroupId', ''), 'status': sg.get('Status', '')}
+        )
+
+    tags = {}
+    if instance.get('TagList'):
+        for tag in instance.get('TagList', []):
+            if 'Key' in tag and 'Value' in tag:
+                tags[tag['Key']] = tag['Value']
+
+    return Instance(
+        instance_id=instance.get('DBInstanceIdentifier', ''),
+        status=instance.get('DBInstanceStatus', ''),
+        engine=instance.get('Engine', ''),
+        engine_version=instance.get('EngineVersion', ''),
+        instance_class=instance.get('DBInstanceClass', ''),
+        endpoint=endpoint,
+        availability_zone=instance.get('AvailabilityZone'),
+        multi_az=instance.get('MultiAZ', False),
+        storage=storage,
+        preferred_backup_window=instance.get('PreferredBackupWindow'),
+        preferred_maintenance_window=instance.get('PreferredMaintenanceWindow'),
+        publicly_accessible=instance.get('PubliclyAccessible', False),
+        vpc_security_groups=vpc_security_groups,
+        db_cluster=instance.get('DBClusterIdentifier'),
+        tags=tags,
+        dbi_resource_id=instance.get('DbiResourceId'),
+        resource_uri=None,
+    )
 
 
 GET_INSTANCE_DETAIL_DOCSTRING = """Get detailed information about a specific Amazon RDS instance.
 
-<use_case>
 Use this resource to retrieve comprehensive details about a specific RDS database instance
-identified by its instance ID. This provides deeper insights than the instance list resource.
-</use_case>
+identified by its instance ID. This provides deeper insights than the instance list resource,
+including configuration, status, endpoints, storage details and more.
 
-<important_notes>
-1. The instance ID must exist in your AWS account and region
-2. The response contains full configuration details about the specified instance
-3. This resource includes information not available in the list view such as storage details,
-   parameter groups, backup configuration, and maintenance windows
-4. Use the instance list resource first to identify valid instance IDs
-5. Error responses will be returned if the instance doesn't exist or there are permission issues
-</important_notes>
+Important notes:
+1. The response contains full configuration details about the specified instance
+2. This resource includes information not available in the list view such as storage details,
+   parameter groups, backup configuration, maintenance windows and security settings
+3. Use the instance list resource first to identify valid instance IDs
+4. Error responses will be returned if the instance doesn't exist or there are permission issues
+5. The response includes all tags associated with the instance
+6. Security group information includes both the ID and current status
 
-## Response structure
-Returns a JSON document containing detailed instance information including:
-- `instance_id`: The unique identifier for the instance
-- `status`: Current status of the instance
-- `engine`: Database engine type
-- `engine_version`: The version of the database engine
-- `endpoint`: Connection endpoint information including address, port and hosted zone
-- `instance_class`: The compute and memory capacity of the instance
-- `availability_zone`: The AZ where the instance is located
-- `multi_az`: Whether the instance is a Multi-AZ deployment
-- `storage`: Detailed storage configuration including type, allocation and encryption status
-- `preferred_backup_window`: When automated backups occur
-- `preferred_maintenance_window`: When maintenance operations can occur
-- `publicly_accessible`: Whether the instance is publicly accessible
-- `vpc_security_groups`: Security groups associated with the instance
-- `db_cluster`: The DB cluster identifier if this instance is part of a cluster
-- `tags`: Any tags associated with the instance
-- `resource_uri`: The full resource URI for this specific instance
+Response fields:
+- instance_id: The unique identifier for the instance
+- status: Current operational status of the instance
+- engine: Database engine type (e.g. mysql, postgres)
+- engine_version: The version of the database engine
+- endpoint: Connection endpoint information including address, port and hosted zone
+- instance_class: The compute and memory capacity of the instance
+- availability_zone: The AZ where the instance is located
+- multi_az: Whether the instance is a Multi-AZ deployment
+- storage: Detailed storage configuration including type, allocation and encryption status
+- preferred_backup_window: When automated backups occur
+- preferred_maintenance_window: When maintenance operations can occur
+- publicly_accessible: Whether the instance is publicly accessible
+- vpc_security_groups: Security groups associated with the instance
+- db_cluster: The DB cluster identifier if this instance is part of a cluster
+- tags: Any tags associated with the instance
+- resource_uri: The full resource URI for this specific instance
 """
 
 
-@mcp.resource(
-    uri='aws-rds://db-instance/{instance_id}',
-    name='GetDBInstanceDetails',
-    mime_type='application/json',
-    description=GET_INSTANCE_DETAIL_DOCSTRING,
-)
 @handle_exceptions
 async def get_instance_detail(
     instance_id: str = Field(..., description='The instance identifier'),
-) -> InstanceModel:
+) -> Instance:
     """Get detailed information about a specific instance as a resource.
 
     Args:
-        instance_id: The instance identifier
+        instance_id: The unique identifier of the RDS instance
 
     Returns:
-        JSON string with instance details
+        Instance: Detailed information about the specified RDS instance
+
+    Raises:
+        ValueError: If the specified instance is not found
     """
     logger.info(f'Getting instance detail resource for {instance_id}')
     rds_client = RDSConnectionManager.get_connection()
